+++ inferelator_ng/design_response_R.py
@@ -4,7 +4,7 @@ Compute design and response by calling R subprocess.
 
 import os
 import pandas as pd
-import utils
+from . import utils
 
 
 DR_module = utils.local_path("R_code", "design_and_response.R")
@@ -39,25 +39,13 @@ def save_R_driver(to_filename, delTmin=0, delTmax=110, tau=45,
     return (to_filename, design_file, response_file)
 
 
-def convert_to_R_df(df):
-    """
-    Convert booleans to "TRUE" and "FALSE" so they will be read correctly from CSV
-    format by R.
-    """
-    new_df = pd.DataFrame(df)
-    for col in new_df:
-        if new_df[col].dtype == 'bool':
-            new_df[col] = [str(x).upper() for x in new_df[col]]
-    return new_df
-
-class DR_driver:
+class DRDriver(utils.RDriver):
 
     """
     Configurable container for calling R subprocess to
     compute design and response.
     """
 
-    target_directory = "/tmp"
     meta_file = "meta_data.csv"
     exp_file = "exp_mat.csv"
     script_file = "run_design_response.R"
@@ -68,11 +56,12 @@ class DR_driver:
     tau = 45
 
     def path(self, filename):
-        return os.path.join(self.target_directory, filename).replace('\\', '/')
+        result = os.path.join(self.target_directory, filename).replace('\\', '/')
+        return utils.r_path(result)
 
     def run(self, expression_data_frame, metadata_dataframe):
-        exp = convert_to_R_df(expression_data_frame)
-        md = convert_to_R_df(metadata_dataframe)
+        exp = utils.convert_to_R_df(expression_data_frame)
+        md = utils.convert_to_R_df(metadata_dataframe)
         exp.to_csv(self.path(self.exp_file))
         md.to_csv(self.path(self.meta_file))
         (driver_path, design_path, response_path) = save_R_driver(
@@ -85,15 +74,10 @@ class DR_driver:
             response_file=self.path(self.response_file),
             design_file=self.path(self.design_file)
         )
-        #subprocess.call(['R', '-f', driver_path])
-        # command = "R -f " + driver_path
-        # execu = 'C:\Program Files\R\R-3.2.2\bin\Rscript.exe'
         utils.call_R(driver_path)
-        # stdout = subprocess.check_output(command, shell=True)
-        # assert stdout.strip().split()[-2:] == [b"done.", b">"], (
-        #     "bad stdout tail: " + repr(stdout.strip().split()[-2:])
-        # )
         final_design = pd.read_csv(design_path, sep='\t')
         final_response = pd.read_csv(response_path, sep='\t')
         return (final_design, final_response)
+
+
